+++ d2/src/main/java/com/linkedin/d2/balancer/ZKFSLoadBalancerWithFacilitiesFactory.java
@@ -107,7 +107,8 @@ public class ZKFSLoadBalancerWithFacilitiesFactory implements LoadBalancerWithFa
                                                    config.serviceDiscoveryEventEmitter,
                                                    config.dualReadStateManager,
                                                    config.loadBalanceStreamException,
-                                                   config.isLiRawD2Client
+                                                   config.isLiRawD2Client,
+                                                   config.enablePotentialClientsCache
     );
   }
 }